Moved jitc_reduce
